Use 16 bit integers for VectorSet index type
Gives another few percent improvement in the metrics I'm measuring

+++ source/blender/blenkernel/BKE_pbvh_api.hh
@@ -98,13 +98,14 @@ class Node {
 struct MeshNode : public Node {
   /**
    * Global vertices are mapped to local indices with a vector set, with a specialized type in
-   * order to use 32 bit integers for slot values.
+   * order to use 32 bit integers for slot values. Use a 16 bit integer for the slot index type
+   * because there will always be less than #leaf_limit vertices in a node.
    */
   using LocalVertMap = VectorSet<int,
                                  DefaultProbingStrategy,
                                  DefaultHash<int>,
                                  DefaultEquality<int>,
-                                 SimpleVectorSetSlot<int, int>,
+                                 SimpleVectorSetSlot<int, int16_t>,
                                  GuardedAllocator>;
 
   /** Indices into the #Mesh::faces() array. Refers to a subset of Tree::prim_indices_. */
